Fixed param destructuring bytecode being out of order
Before we changed the way Parser#parseFunctionParams works, it broke the
destructuring up into base components, and then went through and
assigned individual parameter symbols. When we modified it to use
Parser#variable, we did not keep this call order.

Instead of calling createDestructuringAssignment after we define the
symbol, we need to define all symbols in the primary for loop first.
Then we can go through and call createDestructuringAssignment