Support object literal computed keys
If a computed key is found, it will be stored as a property of the ObjectLiteral,
just like normal keys. However, it is stored as an AstNode, and is later retrieved
to undergo Node transformation and finally Codegen.

Note that as of now, using Symbols as objects keys is still unsupported,
and doing so will result in a nasty crash.