Began work on iterators
The bulk of the task is to transform the generators that are currently
implemented, which are rhino-specific, to ES6-compliant generators.
First and foremost, this includes changing the syntax to use the
asterisk operator, which was easy, and secondly, making sure the return
values align with the spec.

Currently, a generator function will yield all of its values, and when
called again, it will throw an error. This is incorrect behavior; when
the iterator is called after the final yield has been called, the
iterator should return the result object with "value" set to the
function's normal return value (undefined if there is no return, of
course), and "done" set to true.

This commit is an incomplete implementation of the two above
descriptions. Generators will still throw on the final call, however
they do return the correct value/done object (albeit, with incorrect
values for "done").