Fixed pipeline operator w/ arrow functions.
For now, this is a fairly hacky fix. The spec says that the rhs of a pipeline
should parse a LogicalORExpression, which seems to never parse arrow functions.

To work around this issue, I have added a hacky fix that consumes arrow functions manually
inside of a PipelineExpression. This is almost for sure not intended, so we'll have to
investigate this further.