Added operator overloading
This is inspired by
https://github.com/keithamus/ecmascript-operator-overloading-proposal.
Not that there are a couple differences.

First of all, the operator methods are instance methods, not static
methods. So instead of having a static method that accepts two arguments
(or one for unary operators), a class should have an instance method
that accepts one argument (or zero for unary operators).

Second, the operators && and || have been implemented. I see no reason
why these operators cannot be supported.

Lastly, the '+' unary operator does not coerce the return value to a
number. The '-' unary operator is specified to not do this, so it makes
very little sense to have that diverging behavior between the two
operators.