Set const slot value if neither UNINITIALIZED or INITIALIZED flags are present
Inside of a logical statement, such as an if, a call to
ScriptRuntime.enterWith is made, with all necessary symbol keys. If a
symbol is a const, the UNINITIALIZED_CONST flag is not properly set.
However, neither is the INITIALIZED_CONST flag. Therefore, if both of
these flags are not present on a const slot, then we know we should set
the slot value and add the INITIALIZED_CONST flags.