Functions requiring activation objects now correctly apply default arguments
As is described in #8, requiring a function activation meant that
Codegen skipping handling default arguments. This commit fixes the
issue. Additionally, in order to retain correct 'arguments' mapping,
NativeCall accepts two arg arrays: 'callArgs', containing the arguments
the user supplied to the function call, and 'effectiveArgs', the
arguments after applying the default arguments.

Note that the default arguments must be handled in Codegen (and not
delegated to ScriptRuntime) in order to not accidentally throw an
exception if a parameter has a default value set to a throw expression.

Fixes #8