Lexical variables cannot be used before initialization
Previously, Rhino would allow the use of lexical variables before their
initialization. The following code would print 'undefined':

print(foo);
const foo = 10;

This behavior, however, is not correct. V8 throws a ReferenceError, along
with a message about how the user is accessing a lexical variable before
initialization. This commit introduces that behavior to Rhino.

To accomplish this, Slots now have an initialized field. When a slot is
created, it's initialized field is set to false, and is only ever set to
true when Slot#setValue(Object, Scriptable, Scriptable) is called. When
a caller attempt to get the Slot value (via Slot#getValue()), an
exception will be thrown if the Slot's initialized flag is not set.

In order to accommodate this change, two new Scriptable methods have
been added: declare(String, Scriptable) and declareConst(String,
Scriptable). These methods simply insert a new slot directly into the
SlotMap (the latter creating the Slot with an UNINITIALIZED_CONST
attributes), as opposed to the put methods, which associate the new slot
with Undefined.instance. Note that there are no overrides for types int
and Symbol because it is not possible to declare a variable with an int
or Symbol identifier.