Function parameters are considered lexical, and have TDZ detection
NativeFunction#isVarLexical will now return true for function
parameters. Additionally, function parameters now have TDZ detection.
Specifically, the following TDZs are now detected:

  - function(a = a) {}
  - function(a = b, b) {}

These are runtime errors, and thus will compile fine. If such a function
is called, it will throw a ReferenceError stating that 'a' (for the
first example, or 'b', in the case of the second example) is being used
before it has been initialized.