Debug output is now a Context feature
DEBUG_CODEGEN has been replaced with a new Context feature,
Context.EMIT_DEBUG_OUTPUT. This works in function with
setDebugOutputPath, which sets the directory where debug output is
generated. Is the feature is enabled but no file is set, an
IllegalStateException is thrown.

ScriptsRunner has been refactored to use the new debug output feature.
In order to do so, it now programmatically sets up it's environment
instead of just delegating to shell.Main with arguments.