Class FunctionNodes are now associated with the class itself
This fix is primarily concerned with the inability to pass inline
functions to UDDs, such as:

    class Foo {
        @customDecorator(() => {})
        bar() {}
    }

Previously, all FunctionNodes generated inside of a class were attached
to the outer scope (i.e. Codegen#currentScriptOrFn didn't change when
generating the class). This behavior made it impossible to pass an
inline function to a UDD.

The reason for this is that UDDs are handled in three phases: one for
each built in decorator (@wrap, @register, and @initialize). @wrap and
@register work perfectly fine with the old behavior -- they affect the
class itself, and are handled in the upper scope. However, @initialize,
needing to set instance properties, is handled in the constructor. That
means that currentScriptOrFn will be set to that class, >not< the outer
scope, which is what was expected.

This problem aside, not attaching a class's FunctionNodes to that class
is counter-intuitive anyways. Not only does this change fix an important
bug, it also makes class generation much more intuitive.