Runtime: Make private slot implementation much more robust
This gets rid of the whole "toggling private slots" thing in favor of
just passing around an isPrivate boolean to the methods which access
the slotMap (get, put, bunch of codegen methods, etc). This not only
just makes more sense, but avoids the unfortunate side effect where
anything on the right-hand side of a private field assignment couldn't
access any of the normal class slots.